// ssyweb/SSYNEW/src/app/models/boardDetails.model.ts
export class BoardMasterDetails {
    boardId:      number;
    boardName:    string;
    description:  string;
    addressLine1: null;
    addressLine2: null;
    city:         null;
    stateId:      null;
    statusId:     number;
    createdBy:    null;
    creadtedDate: null;
    updatedBy:    null;
    upadtedDate:  null;
}