// ssyweb/SSYNEW/src/app/models/ClaimTypes.model.ts
export class ClaimTypes {
    claimMasterId: number;
    claimName: string;
    cliamEffectiveDate: string;
    claimDescription: null;
    isAnnualLimit: boolean;
    annualLimit: number;
    effectiveDate: string;
    statusId: number;
    createdBy: number;
    createdDate: string;
    updatedBy: number;
    updatedDate: string;
    statusName: string;
}